Make state endpoint request helpers private to remote module
It was previously tempting to either try to use Get/Patch directly as general request helpers, or add unrelated functionality (like we did by adding `make_uri` et al) to the request module.

This moves request as a private submodule of remote, being its sole client/user.

Change-type: patch
